refactor: align stray VOs with canonical base
EpisodeCandidate and EpisodeCompositeId were using @dataclass(frozen=True)
instead of CompoundValueObject. Migrate both:

- EpisodeCandidate inherits CompoundValueObject; defer_build +
  model_rebuild() in entities/__init__ resolves the forward reference
  to WatchProgress without import cycles.
- EpisodeCompositeId moves to shared_kernel/value_objects/; parse()
  now also catches DomainValidationException to preserve the
  None-on-invalid contract.
