+++ src/shared_kernel/value_objects/__init__.py
@@ -1,5 +1,6 @@
 """Shared value objects used across multiple modules."""
 
+from src.shared_kernel.value_objects.episode_composite_id import EpisodeCompositeId
 from src.shared_kernel.value_objects.file_path import FilePath
 from src.shared_kernel.value_objects.image_url import ImageUrl
 from src.shared_kernel.value_objects.language_code import LanguageCode
@@ -11,6 +12,7 @@ from src.shared_kernel.value_objects.user_id import UserId
 __all__ = [
     "AudioTrack",
     "CollectionMediaType",
+    "EpisodeCompositeId",
     "FilePath",
     "ImageUrl",
     "LanguageCode",
